requirements.txt
numpy
pandas
scikit-learn
scikit-image
matplotlib
ipykernel
pooch
requests
Pillow
pydot
tqdm
types-requests
black
flake8
mypy
wandb
tensorflow-addons
tensorflow-io
